update precomputed average edge lengths
H3 reference implementation did the same in https://github.com/uber/h3/pull/749

Couldn't test against it since it's not released yet (and thus
there is no h3ron_h3_sys that binds to it). Not a big deal since I
shouldn't exclusively rely on tests against H3 to ensure my correctness.

Instead I have a little tool (which is equivalent to the Python notebook
linked in the PR) that generated a golden file and I test against it.