geom/polygon: drastically reduce `to_cells` memory usage
Since we're filling rings from the outlines to the center following a
concentric propagation the "seen" set only need to remember the cells of
the previous iteration.

This simple optimization has a HUGE impact on the memory usage.
For instance, when polyfilling at resolution 10:
- France now requires  12.80MB (previously  2.25GB) => ~x180 smaller
- Russia now requires 111.97MB (previously 18.02GB) => ~x165 smaller

Note that the reduced memory usage also give a lil' speedup:
- France went from  13.15s to   8.40s
- Russia went from 559.19s to 405.83s