geom/vertex_graph: improve accuracy and reduce memory usage
This commit adresses two issue

Excessive memory usage
======================

On large inputs, the algorithm consumed way too much memory:
- on France's 44 millions of cells (resolution 10): runtime 2m15s, peak
  memory usage ~13G
- on Russia's 1 billions of cells (resolution 10): dunno, I stopped the
  execution at 79G (and it was still growing...)

This is mainly due to the fact that we are piggybacking the duplicates
detection onto the vertex graph (we identify duplicate cells trhough
triplicated edges).
Because of this approach, we couldn't prune the graph on-the-fly and
(must wait until the complete input has been processed).

Using an online dup check using an HashSet of cells gave the best better
result, but the memory usage was still higher than expected:
- 1.2G for France (way better than the original 13G)
- 35G for Russia (wich is huge considering the input data is around 8G).

This is partly due to memory reallocation scheme and to overallocation
to preserve the hash performance (load factor).
We can avoid the reallocation issue by pre-reserving memory for the
HashSet using the input's size_hint (went from 35G to 26G for Russia)
but it's still not satisfying.

In the end, using a sorted vector (+ dedup pass) is the best
memory-wise:
- France now runs with 627M of RAM
- Russia now runs with 8.7G of RAM
But this is a bit slower than the HashSet approach (but still worth it,
give the huge reduction in memory usage).

I do plan to investigate other approaches later, to see if I can win
back some perf while keeping the RAM usage under control.

Accuracy
========

The code to identifying matching edges is more complicated that it ought
to be. This is due to both floating point error and distortions
(especially around pentagons) which make the matching harder (can't use
strict equality).

This already caused quite a few bugs in the past (see 86e01b57496d and
1dcc3dfff84d), and to be honnest I'm quite unhappy with the current
implementation.

To solve this, I rewrote the algorithm to rely on VertexIndex instead of
geographical ones, this has two big advantages:
- vertices shared between cell are guaranteed to have the same ID ( can
  be matched with a strict equality check)
- a VertexIndex is twice smaller than a LatLng pair (64 vs 128 bit)

The only issue is that vertex index only represent topological vertices
while we also need to insert distortions indexes when its needed (i.e.
cells covering more than one icosahedron faces), but we can add code to
handle that.

The resulting code is simpler (not more custom Hash/Ord implementation)
and a bit faster:
- France went from 1m38s to 1m9s
- Russia went from 48m22s to 40m3s