geom: add support for more polyfill modes
Add support for:
- ContainsCentroid: this was the existing mode
- IntersectsBoundary: to generate set that fully cover the given shape
- ContainsBoundary: because why not (was mentioned here [1])

For the `contains` predicate we need to wrap the ring (a `LineString`)
into a `Polygon` to get the right behavior.
As a consequence, `Ring` always takes the ownership of its input, and
I've propagated that to every shape to simplify the API.

As expected, these new modes are slower than the old one since checking
if a shape intersects/is within another is more costly than checking a
single point.

For Paris at resolution 10:
- ContainsCentroid:   0.19s for 7 834 cells
- IntersectsBoundary: 0.29s for 8 112 cells
- ContainsBoundary:   0.32s for 7 547 cells

For France at resolution 10:
- ContainsCentroid:     7.29s for 44 250 550 cells
- IntersectsBoundary: 173.14s for 44 297 193 cells
- ContainsBoundary:   ??? (still running after 18h...)

There is still work to do on ContainsBoundary as it seems to blow up at
larger scale...

[1]: https://github.com/uber/h3/blob/master/dev-docs/RFCs/v4.0.0/polyfill-modes-rfc.md