geom/to_h3: fix transmeridian cells on non-transmeridian shapes
The code was only translating cell boundaries when working against
transmeridian shapes.

But when testing a transmeridian cell against a non-transmeridian shape,
you also needs to translate the cell boundaries, otherwise some edges
are going "backward" and may generate false positive, which derails the
algorithm.

Took the occasion to refactor some duplicated code and to use a Ring to
represents the cell boundaries, which allows to "cache" the bounding box
(and handle the transmeridian translation) when checking a single cell
against multiple shapes, e.g. the rings of a (multi)-polygon.

Closes: #21