geom/ring: improve transmeridian handling
Try to reduce the number of false positives by checking if the "fixed"
shape is valid.
It's not bullet-proof but should improve the situation.

What was happening in #23 is that the antimeridian heuristic was wrongly
triggered, which resulted in an invalid shape (self-intersecting
polygon) and from there the polyfill algorithm went astray.

Closes: #23