geom/to_h3: fix transmeridian cells with western non-transmeridian shapes
d3925b16e93d fixed the issue for eastern shapes (at the left of the
antimeridian), but we still had an issue with western shapes (at the
right of the antimeridian).

The problem is that we're handling transmeridian geometries by
translating them to the right and while this does the trick for shape at
the left of the transmeridian, it's doesn't for shape at the right (and
vice versa).

So now, when checking containment predicate for transmeridian cells, we
check against both translations.

Seems like this bugfix added a bunch of cell that we were missing in the
issue_23 tests (hard to say which ones are legit or not since they cover
the pole singularity)