geom/to_geom: fix wrong application of transmeridian handling
The coordinates correction for transmeridian shapes was always applied,
which triggered some issues, like polygon being considered within another
even though it wasn't the case, for shape crossing the prime meridian
(which doesn't need any adjustment).
And since we were testing a single point + the order of the point were
not deterministic due to the usage of hashmap, the bug was not always
visible.