geom: optimize Tiler by using PreparedGeometry
Performance on ContainsCentroid are unchanged as we keep the existing
logic (the `relate` approach was a bit slower in this case), but for the
other predicates we get a x15-x17 speedup thanks to the
PreparedGeometry!!!
                                          OLD              NEW
    Mode/Centroid/Full                 13.7±0.03ms      13.6±0.03ms
    Mode/Centroid/Holes                14.9±0.13ms      14.6±0.03ms
    Mode/Centroid/Transmeridian        10.0±0.02ms      10.1±0.02ms
    Mode/Contains/Full                801.4±9.55ms      46.5±0.14ms
    Mode/Contains/Holes                45.8±2.58ms      44.6±0.24ms
    Mode/Contains/Transmeridian        62.2±5.29ms      28.5±0.36ms
    Mode/Intersects/Full             716.1±14.00ms      46.3±0.12ms
    Mode/Intersects/Holes              40.4±0.25ms      44.6±0.10ms
    Mode/Intersects/Transmeridian      53.1±0.28ms      28.3±0.17ms

And for France at resolution 10 (44 millions of cells):
- ContainsCentrois:     7.2s =>  7.0s
- IntersectsBoundary: 173.1s => 10.9s
- ContainsBoundary:   203.9s => 10.6s

In term of unit tests, the new handling of the transmeridian shapes
changed a bit the estimation of the coverage size.
In some case we overestimate more (we do worse) and and in some case we
overestimate less (we do better).
Since this function is mainly here for H3 compatibility (not really
useful in Rust) and considering that H3 itself now have a new estimator
(that I should probably port someday as it is aware of the containement
mode) it's not a blocker if I'm not iso-functional on that one for now.