geom/tiler: fix boundary stitching around transmeridian
We used to take the closest value on the "right" side of the
antimeridian as the value to fix. But seems like with the latest update
of i-overlay, the rounding error can go into the other direction and the
value to fix may be on the "wrong" side of the antimeridian.

The stitching logic was then fixing the wrong point, resulting in
invalid shape that would trip a debug assert in the geo code.

To fix the issue we're now looking at both sides, using an epsilon.
It's still pretty shady and brittle so I'm not really happy about it,
but I can't come up with anything better for now.