geom: implement heterogeneous set dissolution
This is way faster for very large and dense set of cells (e.g. cities or
countries) and it uses less memory as well.
That being said, it seems to be slower for simpler case so we also keep
the old approach.

                    Homogeneous     Heterogeneous
    Dissolve/6        6.4±0.07µs       6.3±0.09µs
    Dissolve/7       48.8±0.82µs      61.5±1.27µs
    Dissolve/8      240.9±3.22µs     407.9±4.46µs
    Dissolve/9        2.0±0.03ms       3.6±0.02ms
    Dissolve/10      12.8±0.35ms      18.4±0.23ms
    Dissolve/11     119.3±7.98ms      99.2±2.80ms
    Dissolve/12    946.1±33.75ms     400.3±2.88ms

And without duplicate detection:

                    Homogeneous     Heterogeneous
    Dissolve/6        6.1±0.04µs       6.1±0.06µs
    Dissolve/7       46.2±0.32µs      60.4±0.93µs
    Dissolve/8      228.9±1.50µs     401.6±2.85µs
    Dissolve/9        1.9±0.02ms       3.6±0.10ms
    Dissolve/10      12.7±0.59ms      18.0±0.17ms
    Dissolve/11    123.2±10.46ms      97.7±2.62ms
    Dissolve/12    919.8±34.98ms     389.5±3.72ms

Working at larger scale, the benefits are even more visible!!!
For example, testing on H3 cell sets that conver Switzerland at various
resolutions we have the following results (timing in seconds):

    resolution          Uber        homogeneous    heterogeneous
     (#cells)            H3            h3o              h3o
     1 (1)              0.000008     0.000016        0.000017
     2 (4)              0.000017     0.000015        0.000017
     3 (11)             0.000034     0.000024        0.000025
     4 (44)             0.000102     0.000059        0.000096
     5 (220)            0.000419     0.000274        0.000504
     6 (1,334)          0.002306     0.001439        0.002750
     7 (8,701)          0.014725     0.011644        0.015733
     8 (59,154)         0.105578     0.093431        0.077385
     9 (409,207)        1.057587     0.905263        0.380793
    10 (2,851,362)    421.549715     6.559264        1.738701
    11 (19,924,463)     N/A         65.077335        9.705693

The Uber H3 implementation was still running after 4h, so I've stopped
it.

In term of RAM usage, the heterogeneous is also a huge win: for example
at resolution 11, we go from 3.77GiB (3.73GiB if we skip the dup
detection) to 432MiB (20MiB if we skip the dup detection).

Since we now have a bunch of toggles for the dissolution algorithm,
let's replace the free function by the builder pattern. This will be
more future-proof anyway.

(cherry picked from commit 9f19e600448fef3a813015be4c6de46ec36a2d4a)