rewrite compact to work in-place
Current compaction API has always been kinda misleading: taking an
iterator as input and producing an iterator as output you would expect
this function to be fully lazy.
But due to how the dup & homogeneity check is implemented, this actually
trigger an eager consumption of the iterator and the allocation of a new
vector: so much for lazyness...

By exposing an API that work in-place we make this explicit to the
caller.
We do lose the lazyness of the output, but we're able to avoid
allocation altogether by updating the input itself, so all in all it's
still a win (+ I haven't come accross a case where we only want to
partially compact the input).

A little 30% speedup:

    bench              h3o-old        h3o-new
    -----              -----------    -----------
    FullCompaction     91.2±15.43µs    70.4±6.40µs
    NoCompaction       202.4±14.62µs  151.7±28.20µs
    PartialCompaction  171.3±0.62µs   136.2±17.18µs

The old API can be reproduced in 3 lines if needed:

    let mut cells = input.collect::<Vec<_>>();
    CellIndex.compact(&mut cells)?;
    return cells.into_iter();